Changed Datatype

+++ ProblemH.java
@@ -4,12 +4,12 @@ public class ProblemH {
     public static void main(String[] args) {
         Scanner hr = new Scanner(System.in);
 
-        int a = hr.nextInt(), b = hr.nextInt();
+        long a = hr.nextLong(), b = hr.nextLong();
 
         double result = (double) a / b;
     
-        System.out.println("floor " + a + " / " + b + " = " + (int)Math.floor(result));
-        System.out.println("ceil " + a + " / " + b + " = " + (int)Math.ceil(result));
+        System.out.println("floor " + a + " / " + b + " = " + (long)Math.floor(result));
+        System.out.println("ceil " + a + " / " + b + " = " + (long)Math.ceil(result));
         System.out.println("floor " + a + " / " + b + " = " + Math.round(result));
 
         hr.close();
